style: introduce MFL Actions
